pyproject.toml
[project]
name = "chat-with-docs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "numpy>=2.4.2",
    "openai>=2.16.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pylint>=4.0.4",
    "pytest>=9.0.2",
    "ruff>=0.15.0",
]
